+++ wgpu-hal/src/dx12/types.rs
@@ -32,6 +32,13 @@ winapi::ENUM! {
     }
 }
 
+winapi::ENUM! {
+    enum D3D12_WAVE_MMA_TIER  {
+        D3D12_WAVE_MMA_TIER_NOT_SUPPORTED = 0,
+        D3D12_WAVE_MMA_TIER_1_0 = 10,
+    }
+}
+
 winapi::STRUCT! {
     struct D3D12_FEATURE_DATA_D3D12_OPTIONS3 {
         CopyQueueTimestampQueriesSupported: winapi::shared::minwindef::BOOL,
@@ -39,5 +46,38 @@ winapi::STRUCT! {
         WriteBufferImmediateSupportFlags: D3D12_COMMAND_LIST_SUPPORT_FLAGS,
         ViewInstancingTier: D3D12_VIEW_INSTANCING_TIER,
         BarycentricsSupported: winapi::shared::minwindef::BOOL,
+    }
+}
+
+winapi::STRUCT! {
+    struct D3D12_FEATURE_DATA_D3D12_OPTIONS9 {
+        MeshShaderPipelineStatsSupported: winapi::shared::minwindef::BOOL,
+        MeshShaderSupportsFullRangeRenderTargetArrayIndex: winapi::shared::minwindef::BOOL,
+        AtomicInt64OnTypedResourceSupported: winapi::shared::minwindef::BOOL,
+        AtomicInt64OnGroupSharedSupported: winapi::shared::minwindef::BOOL,
+        DerivativesInMeshAndAmplificationShadersSupported: winapi::shared::minwindef::BOOL,
+        WaveMMATier: D3D12_WAVE_MMA_TIER,
+    }
+}
+
+winapi::ENUM! {
+    enum D3D_SHADER_MODEL {
+        D3D_SHADER_MODEL_NONE = 0,
+        D3D_SHADER_MODEL_5_1 = 0x51,
+        D3D_SHADER_MODEL_6_0 = 0x60,
+        D3D_SHADER_MODEL_6_1 = 0x61,
+        D3D_SHADER_MODEL_6_2 = 0x62,
+        D3D_SHADER_MODEL_6_3 = 0x63,
+        D3D_SHADER_MODEL_6_4 = 0x64,
+        D3D_SHADER_MODEL_6_5 = 0x65,
+        D3D_SHADER_MODEL_6_6 = 0x66,
+        D3D_SHADER_MODEL_6_7 = 0x67,
+        D3D_HIGHEST_SHADER_MODEL = 0x67,
+    }
+}
+
+winapi::STRUCT! {
+    struct D3D12_FEATURE_DATA_SHADER_MODEL {
+        HighestShaderModel: D3D_SHADER_MODEL,
     }
 }