+++ wgpu-hal/src/vulkan/adapter.rs
@@ -41,6 +41,7 @@ pub struct PhysicalDeviceFeatures {
     ray_query: Option<vk::PhysicalDeviceRayQueryFeaturesKHR>,
     zero_initialize_workgroup_memory:
         Option<vk::PhysicalDeviceZeroInitializeWorkgroupMemoryFeatures>,
+    shader_int64_atomic: Option<vk::PhysicalDeviceShaderAtomicInt64Features>,
 }
 
 // This is safe because the structs have `p_next: *mut c_void`, which we null out/never read.
@@ -86,6 +87,9 @@ impl PhysicalDeviceFeatures {
             info = info.push_next(feature);
         }
         if let Some(ref mut feature) = self.ray_query {
+            info = info.push_next(feature);
+        }
+        if let Some(ref mut feature) = self.shader_int64_atomic {
             info = info.push_next(feature);
         }
         info
@@ -351,6 +355,17 @@ impl PhysicalDeviceFeatures {
             } else {
                 None
             },
+            shader_int64_atomic: if device_api_version >= vk::API_VERSION_1_1
+                || enabled_extensions.contains(&vk::KhrGetPhysicalDeviceProperties2Fn::name())
+            {
+                Some(
+                    vk::PhysicalDeviceShaderAtomicInt64Features::builder()
+                        .shader_buffer_int64_atomics(true)
+                        .build(),
+                )
+            } else {
+                None
+            },
         }
     }
 
@@ -471,6 +486,11 @@ impl PhysicalDeviceFeatures {
         features.set(F::SHADER_F64, self.core.shader_float64 != 0);
         features.set(F::SHADER_INT64, self.core.shader_int64 != 0);
         features.set(F::SHADER_I16, self.core.shader_int16 != 0);
+
+        features.set(
+            F::SHADER_INT64_ATOMIC,
+            caps.supports_extension(vk::KhrShaderAtomicInt64Fn::name()),
+        );
 
         //if caps.supports_extension(vk::KhrSamplerMirrorClampToEdgeFn::name()) {
         //if caps.supports_extension(vk::ExtSamplerFilterMinmaxFn::name()) {
@@ -806,6 +826,11 @@ impl PhysicalDeviceCapabilities {
         // Require `VK_EXT_texture_compression_astc_hdr` if the associated feature was requested
         if requested_features.contains(wgt::Features::TEXTURE_COMPRESSION_ASTC_HDR) {
             extensions.push(vk::ExtTextureCompressionAstcHdrFn::name());
+        }
+
+        // Require `VK_EXT_shader_image_atomic_int64` if the associated feature was requested
+        if requested_features.contains(wgt::Features::SHADER_INT64_ATOMIC) {
+            extensions.push(vk::KhrShaderAtomicInt64Fn::name());
         }
 
         extensions
@@ -1456,6 +1481,10 @@ impl super::Adapter {
 
             if features.contains(wgt::Features::SHADER_INT64) {
                 capabilities.push(spv::Capability::Int64);
+            }
+
+            if features.contains(wgt::Features::SHADER_INT64_ATOMIC) {
+                capabilities.push(spv::Capability::Int64Atomics);
             }
 
             let mut flags = spv::WriterFlags::empty();